fix(mq): move match media utils functions to lib
